add next button to structural search tooltip
